# api/rest/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Service for geo location by IP based on GeoLite2 database",
        "title": "Geos API",
        "contact": {},
        "license": {
            "name": "ᅠ"
        },
        "version": "1.0"
    },
    "basePath": "/geoip",
    "paths": {
...
        "/geoname/city": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geonames"
                ],
                "summary": "city",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "comma separated list of country codes",
                        "name": "country-codes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Geoname"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/geoname/country": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geonames"
                ],
                "summary": "country",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "comma separated list of country codes",
                        "name": "country-codes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.GeoNameCountry"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/geoname/subdivision": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geonames"
                ],
                "summary": "city lite",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "comma separated list of country codes",
                        "name": "country-codes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.AdminSubdivision"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.AdminSubdivision": {
            "type": "object",
            "properties": {
                "asciiName": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "geoNameID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "entity.GeoNameCountry": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "number"
                },
                "capital": {
                    "type": "string"
                },
                "continent": {
                    "type": "string"
                },
                "currencyCode": {
                    "type": "string"
                },
                "currencyName": {
                    "type": "string"
                },
                "equivalentFipsCode": {
                    "type": "string"
                },
                "fips": {
                    "type": "string"
                },
                "geoNameID": {
                    "type": "integer"
                },
                "iso2Code": {
                    "type": "string"
                },
                "iso3Code": {
                    "type": "string"
                },
                "isoNumeric": {
                    "type": "string"
                },
                "languages": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "neighbours": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "population": {
                    "type": "integer"
                },
                "postalCodeFormat": {
                    "type": "string"
                },
                "postalCodeRegex": {
                    "type": "string"
                },
                "tld": {
                    "type": "string"
                }
            }
        },
        "entity.Geoname": {
            "type": "object",
            "properties": {
                "admin1Code": {
                    "type": "string"
                },
                "admin2Code": {
                    "type": "string"
                },
                "admin3Code": {
                    "type": "string"
                },
                "admin4Code": {
                    "type": "string"
                },
                "alternateCountryCodes": {
                    "type": "string"
                },
                "alternateNames": {
                    "type": "string"
                },
                "asciiName": {
                    "type": "string"
                },
                "class": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "countryCode": {
                    "type": "string"
                },
                "dem": {
                    "type": "integer"
                },
                "elevation": {
                    "type": "integer"
                },
                "geoNameID": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "population": {
                    "type": "integer"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
    }
}